// app/src/backend-client/api.ts
export interface IntervalA {
    /**
     * 
     * @type {number}
     * @memberof IntervalA
     */
    'start': number;
    /**
     * 
     * @type {number}
     * @memberof IntervalA
     */
    'end': number;
}
/**
 * 
 * @export
 * @interface PlantProfile
 */
export interface PlantProfile {
    /**
     * 
     * @type {string}
     * @memberof PlantProfile
     */
    'name': string;
    /**
     * 
     * @type {PlantProfileVariables}
     * @memberof PlantProfile
     */
    'variables': PlantProfileVariables;
}
/**
 * 
 * @export
 * @interface PlantProfileVariables
 */
export interface PlantProfileVariables {
    /**
     * 
     * @type {IntervalA}
     * @memberof PlantProfileVariables
     */
    'lightMmol': IntervalA;
    /**
     * 
     * @type {IntervalA}
     * @memberof PlantProfileVariables
     */
    'lightLux': IntervalA;
    /**
     * 
     * @type {IntervalA}
     * @memberof PlantProfileVariables
     */
    'temperature': IntervalA;
    /**
     * 
     * @type {IntervalA}
     * @memberof PlantProfileVariables
     */
    'humidity': IntervalA;
    /**
     * 
     * @type {IntervalA}
     * @memberof PlantProfileVariables
     */
    'soilMoisture': IntervalA;
    /**
     * 
     * @type {IntervalA}
     * @memberof PlantProfileVariables
     */
    'soilEc': IntervalA;
}